Updated cli and html-to-pdf stuff
Made some changes to the html-to-pdf to run tests locally; /bin/sh doesn't have pipefail, that's a bash thing.
Also made some changes to the real code. Currently it will try to save to s3 AND return the base64 encoded version of
the pdf. One or the other will be sufficient here; either the web app can save to s3, or the lambda.