Added cf template, and fixed lambda code
Lambda takes in html as a string (maybe this should be base64d?), converts the string to pdf, and saves it out to s3.
Cloudformation makes the lambda and its role, and the role's policy.
I imagine this will take care of making the s3 bucket as well at some point, but i ran into a chicken egg problem trying to
create the lambda function with an s3 zip file. Maybe we have an s3 bucket for lambda functions that cf templates can pull from?